+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the Rundeck cookbook.
 
 ## Unreleased
 
+- Pointing to a valid repository, as Bintray has been depricated.
+
 ## 5.2.2 - *2021-08-30*
 
 - Standardise files with files in sous-chefs/repo-management